used lu and added data type

+++ search_algorithms/0-linear.c
@@ -17,7 +17,7 @@ int linear_search(int *array, size_t size, int value)
 	while (i < size)
 {
 		i++;
-		printf("Value checked array[%zu] = [%d]\n", i, array[i]);
+		printf("Value checked array[%lu] = [%d]\n", (unsigned long)i, array[i]);
 		if (array[i] == value)
 {
 			return (i);
